+++ src/app/app.strings.ts
@@ -10,7 +10,8 @@ export const AppTranslations = {
         'EXAMPLES.FOCUSABLE_ELEMENT': 'Focusable element',
         'EXAMPLES.SELECTED_DESCRIPTION': 'allows to navigate over non-focuseable elements using arrow keys',
         'EXAMPLES.INFINITE_SCROLL_DESCRIPTION': 'allows to upload data in chunks while user is scrolling',
-        'EXAMPLES.DRAGGABLE_DESCRIPTION': 'implements drag & drop functionality'
+        'EXAMPLES.DRAGGABLE_DESCRIPTION': 'implements drag & drop functionality',
+        'EXAMPLES.SELECTED.DISABLED': 'Disable selection'
     },
     ru: {
         'en': 'English',
@@ -22,6 +23,7 @@ export const AppTranslations = {
         'EXAMPLES.FOCUSABLE_ELEMENT': 'Фокусируемый элемент',
         'EXAMPLES.SELECTED_DESCRIPTION': 'позволяет перемещаться по нефокусируемым элементам управления с помощью клавиш со стрелками',
         'EXAMPLES.INFINITE_SCROLL_DESCRIPTION': 'позволяет загружать данные частями во время прокрутки пользователя',
-        'EXAMPLES.DRAGGABLE_DESCRIPTION': 'реализация функции перетаскивания (drag & drop)'
+        'EXAMPLES.DRAGGABLE_DESCRIPTION': 'реализация функции перетаскивания (drag & drop)',
+        'EXAMPLES.SELECTED.DISABLED': 'Запретить смену элемента'
     }
 };