+++ src/app/selected-example/selected-example.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MatListModule } from '@angular/material';
+import { MatListModule, MatCheckboxModule } from '@angular/material';
 import { TranslateModule } from '@ngx-translate/core';
 import { PipSelectedModule } from 'pip-webui2-behaviors';
 
@@ -10,6 +10,7 @@ import { SelectedExampleComponent } from './selected-example.component';
   imports: [
     CommonModule,
     MatListModule,
+    MatCheckboxModule,
     TranslateModule.forChild(),
 
     PipSelectedModule